chore: move routing logic and share code loading to the backend
