+++ src/utils/rpc.ts
@@ -1,4 +1,4 @@
-import type { ApiType } from "@/server/api";
+import type { ApiType } from "@/server/routes/api";
 import { hc } from "hono/client";
 
 export const rpc = hc<ApiType>("/api");
